+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - `REMP` supported. `ProcessingEvent` enum is extended with `REMP` statuses (enum of events posted into `processing.wait_for_transaction` function callback )
 - supported changed interface of ton_types::Cell
 - supported removing Copy interface from UInt256
+- UNSTABLE. `first_remp_status_timeout` and `next_remp_status_timeout` parameters in network config
 
 ## [1.32.0] – 2022-03-22